requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.1.0
drf-yasg==1.21.9
gunicorn==23.0.0
